Linked List
This is a linked list (only forwards). It is possible to do a variety
of functions to the list. It currently only supports int in its
structure, but I plan to improve on it. I also would like to improve
the algorithm that handles insertion at a specific location. I did not
handle for inserting at the end of the list originally, and therefore I
paid the price of having to create many worthless and ugly variables.